requirements.txt
Django
djangorestframework
PyJWT
psycopg2-binary
python-dotenv
django-cors-headers
djangorestframework-csv